Place `StatusExt` and `RpcStatusExt` into separate files.
Also does some minor module imports cleanup.

The goal is to make room for a `CodeExt` trait.
