+++ tonic-types/src/lib.rs
@@ -182,12 +182,7 @@ pub mod pb {
 pub use pb::Status;
 
 mod richer_error;
-
-pub use richer_error::{
-    BadRequest, DebugInfo, ErrorDetail, ErrorDetails, ErrorInfo, FieldViolation, Help, HelpLink,
-    LocalizedMessage, PreconditionFailure, PreconditionViolation, QuotaFailure, QuotaViolation,
-    RequestInfo, ResourceInfo, RetryInfo, RpcStatusExt, StatusExt,
-};
+pub use richer_error::*;
 
 mod sealed {
     pub trait Sealed {}
