+++ tonic-types/src/richer_error/error_details/mod.rs
@@ -1,12 +1,8 @@
+pub(super) mod vec;
+
 use std::{collections::HashMap, time};
 
-use super::std_messages::{
-    BadRequest, DebugInfo, ErrorInfo, FieldViolation, Help, HelpLink, LocalizedMessage,
-    PreconditionFailure, PreconditionViolation, QuotaFailure, QuotaViolation, RequestInfo,
-    ResourceInfo, RetryInfo,
-};
-
-pub(crate) mod vec;
+use super::std_messages::*;
 
 /// Groups the standard error messages structs. Provides associated
 /// functions and methods to setup and edit each error message independently.
