+++ tonic-types/src/richer_error/error_details/vec.rs
@@ -1,7 +1,4 @@
-use super::super::std_messages::{
-    BadRequest, DebugInfo, ErrorInfo, Help, LocalizedMessage, PreconditionFailure, QuotaFailure,
-    RequestInfo, ResourceInfo, RetryInfo,
-};
+use super::super::std_messages::*;
 
 /// Wraps the structs corresponding to the standard error messages, allowing
 /// the implementation and handling of vectors containing any of them.
