+++ tonic-types/src/richer_error/std_messages/error_info.rs
@@ -3,9 +3,9 @@ use std::collections::HashMap;
 use prost::{DecodeError, Message};
 use prost_types::Any;
 
-use crate::richer_error::FromAnyRef;
+use crate::pb;
 
-use super::super::{pb, FromAny, IntoAny};
+use super::super::*;
 
 /// Used to encode/decode the `ErrorInfo` standard error message described in
 /// [error_details.proto]. Describes the cause of the error with structured
