+++ tonic-types/src/richer_error/std_messages/help.rs
@@ -1,9 +1,9 @@
 use prost::{DecodeError, Message};
 use prost_types::Any;
 
-use crate::richer_error::FromAnyRef;
+use crate::pb;
 
-use super::super::{pb, FromAny, IntoAny};
+use super::super::*;
 
 /// Used at the `links` field of the [`Help`] struct. Describes a URL link.
 #[derive(Clone, Debug)]
