+++ tonic-types/src/richer_error/std_messages/mod.rs
@@ -1,39 +1,29 @@
 mod retry_info;
-
 pub use retry_info::RetryInfo;
 
 mod debug_info;
-
 pub use debug_info::DebugInfo;
 
 mod quota_failure;
-
 pub use quota_failure::{QuotaFailure, QuotaViolation};
 
 mod error_info;
-
 pub use error_info::ErrorInfo;
 
 mod prec_failure;
-
 pub use prec_failure::{PreconditionFailure, PreconditionViolation};
 
 mod bad_request;
-
 pub use bad_request::{BadRequest, FieldViolation};
 
 mod request_info;
-
 pub use request_info::RequestInfo;
 
 mod resource_info;
-
 pub use resource_info::ResourceInfo;
 
 mod help;
-
 pub use help::{Help, HelpLink};
 
 mod loc_message;
-
 pub use loc_message::LocalizedMessage;
