+++ tonic-types/src/richer_error/std_messages/quota_failure.rs
@@ -1,9 +1,9 @@
 use prost::{DecodeError, Message};
 use prost_types::Any;
 
-use crate::richer_error::FromAnyRef;
+use crate::pb;
 
-use super::super::{pb, FromAny, IntoAny};
+use super::super::*;
 
 /// Used at the `violations` field of the [`QuotaFailure`] struct. Describes a
 /// single quota violation.
